+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.1.0] - 2023-07-26
+
+- Changed User-Agent header to custom value to differentiate SDKs in the backend
+- Exposed custom audio format and sample rate settings for TTS Streaming
+
 ## [v1.0.2] - 2023-06-07
 
 - Fixed error handling when user passes invalid VoiceID in the TTS Streaming